Refactor: merge clustertemplate and template
