+++ pkg/kapis/devops/v1alpha1/register.go
@@ -20,7 +20,6 @@ import (
 	"kubesphere.io/devops/pkg/api/devops/v1alpha1"
 	"kubesphere.io/devops/pkg/apiserver/runtime"
 	"kubesphere.io/devops/pkg/kapis/devops/v1alpha1/argocd"
-	"kubesphere.io/devops/pkg/kapis/devops/v1alpha1/clustertemplate"
 	"kubesphere.io/devops/pkg/kapis/devops/v1alpha1/common"
 	"kubesphere.io/devops/pkg/kapis/devops/v1alpha1/template"
 )
@@ -34,7 +33,6 @@ func AddToContainer(container *restful.Container, options *common.Options) []*re
 	for _, service := range services {
 		template.RegisterRoutes(service, options)
 		argocd.RegisterRoutes(service, options)
-		clustertemplate.RegisterRoutes(service, options)
 		container.Add(service)
 	}
 	return services
